Update ShipPlanner.java
Updated line 217, this is to address the error where a player can start the game without placing their last ship. This error was replicable by using either mouse button to click on an existing ship or its surroundings. These placements are invalid, but you could press the Ok button afterward. I fixed this by adding an additional "isValidPosition" check whenever a press is done, and only progressing forward if the click is in a valid location.