Update ShipPlanner.java (#37)
This push is to fix the error that a player can start the game without placing their last ship. This is done by clicking anywhere a ship cannot be placed. I fixed this by adding a check in line 217, checking if the mouseclick is a valid location, and only recording the action if it is valid.
Closes #8