Update GameBoard.java
Addressed the issue where the last successful hit wouldn't change to the correct color, turning orange instead of red. This was resolved by manually setting the color of the last position to red before displaying the message. In the original code, when someone got a correct hit, the tile would turn red. The problem stemmed from this code turning the tile red wouldn't occur when someone won due to the if-else structure of this part of the code. As such, manually setting the tile to red when displaying the victory screen resolves this issue.